requirements.txt
agent-framework-core>=1.0.0rc1
azure-identity>=1.19.0
httpx>=0.27.0
python-dotenv>=1.0.0
